requirements.txt
pytest
pytest-selenium
allure-python-commons